requirements.txt
flask==2.3.3
scrapy==2.11.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
groq==0.4.2
python-dotenv==1.0.0  